requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
qdrant-client==1.11.0
sentence-transformers==3.1.1
torch==2.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
PyYAML==6.0.2
openai==1.51.0
numpy==1.26.4
